added delete functionality to admin page and --watch on npm build script command
