+++ package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "start": "node server.js",
     "dev": "vite",
-    "build": "vite build",
+    "build": "vite build --watch",
     "lint": "eslint . --ext js,jsx --report-unused-disable-directives --max-warnings 0",
     "preview": "vite preview",
     "test": "jest --watchAll --config=jest.config.cjs",
@@ -53,7 +53,8 @@
     "jest": "^29.7.0",
     "jest-environment-jsdom": "^29.7.0",
     "prettier": "^3.3.0",
-    "vite": "^5.2.0"
+    "vite": "^5.2.0",
+    "watch": "^1.0.2"
   },
   "extends": [
     "react-app",
